scraping/scraper/requirements.txt
beautifulsoup4==4.6.0
certifi==2018.4.16
chardet==3.0.4
idna==2.7
numpy==1.14.5
Pillow==5.2.0
pytesseract==0.2.2
requests==2.19.1
urllib3==1.23
pysocks==1.6.8
boto3==1.7.52
datadog==0.22.0
user_agent==0.1.9
